# swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Henohenoapi
  description: >-
    The project aims to create a platform for managing exams, teachers can
    public exams(which have a deadline) and add tasks to them, while students
    can submit their answers and answer.
#dominio e baseroot
host: henoheno-app.herokuapp.com
basePath: /
...
schemes:
  - http
#formati dati accettabile da server
consumes:
  - application/json
#formati dati prodotto da server
produces:
  - application/json
...
security:
  - Api_key_student: []     #apikey di studente
  - Api_key_teacher: []     #apikey di insegnante



#parte di url--------------------------------------------------------------------------
paths:
...
  /users:
    get:
      summary: users list
      description: get a users list
      tags:
                - Users_resource
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Users_in_response'
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'
        
          
    post:
      summary: user subscription
      description: create a new user
      tags:
        - Users_resource
      security:
              - {} #non usare nessun api-key
      parameters:
        - $ref: '#/parameters/User_in_post'
      responses:
        '201':
                description: Created
                schema:
                        type: integer
                        description: the new user 's id
                        
        '204':
                $ref: '#/responses/204_no_content'
        '400':
                $ref: '#/responses/400_bad_request'
        '401':
                $ref: '#/responses/401_unauthorized'
        '500':
                $ref: '#/responses/500_internal_server_error'
  '/users/{user_id}':
    get:
      summary: access one specific user
      description: access one specific user
      tags:
        - Users_resource
      parameters:
        - $ref: '#/parameters/User_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User_in_response'
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'
  '/users/{user_id}/exams':
    get:
      summary: get the exams assigned to a student
      description: get the exams assigned to a student
      tags:
        - Users_resource
      
      parameters:
        - $ref: '#/parameters/User_id_in_path'
        - in: query
          name: selection
          schema:
            type: string
          description: Assumes values "created" or "assigned"
      responses:
                '200':
                        description: OK
                        schema:
                              $ref: '#/definitions/Exams'  
                '401':
                        $ref: '#/responses/401_unauthorized'
                '404':
                        $ref: '#/responses/404_not_found'
                '500':
                        $ref: '#/responses/500_internal_server_error'
  '/users/{user_id}/exam_submissions':
    get:
      summary: get the submissions of a student
      description: get the submissions of a student
      tags:
        - Users_resource
      security:
        - Api_key_teacher: []  #solo insegnante
      parameters:
        - $ref: '#/parameters/User_id_in_path'
      responses:
                '200':
                        description: OK
                        schema:
                              $ref: '#/definitions/Exam_submissions' 
                '401':
                        $ref: '#/responses/401_unauthorized'
                '404':
                        $ref: '#/responses/404_not_found'
                '500':
                        $ref: '#/responses/500_internal_server_error'
...
  /groups:
    get:
      summary: groups list
      description: get the list of created groups or the list of groups you're in
      tags:
        - Groups_resource
      responses:
        '200':
          description: OK
          schema:
              $ref: '#/definitions/Groups'
                
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'

        
          
    post:
      summary: create new group
      description: create new group
      tags:
        - Groups_resource
      security:
        - Api_key_teacher: []     #solo insegnante 
      parameters:
        - $ref: '#/parameters/Group_in_post'
      responses:
        '201':
                description: Created
                schema:
                        type: integer
                        description: the new group 's id
        '204':
                $ref: '#/responses/204_no_content'
        '400':
                $ref: '#/responses/400_bad_request'
        '401':
                $ref: '#/responses/401_unauthorized'
        '500':
                $ref: '#/responses/500_internal_server_error'
  '/groups/{group_id}':
    get:
      summary: see the members of a group
      description: get a specific group
      tags:
        - Groups_resource
      parameters:
        - $ref: '#/parameters/Group_id_in_path'
      responses:
        '200':
          description: OK
          schema:
              $ref: '#/definitions/Groups'
              
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'
    put:
      summary: update an existing group
      description: update an existing group
      tags:
        - Groups_resource
      security:
        - Api_key_teacher: []     #solo insegnante 
      parameters:
        - $ref: '#/parameters/Group_id_in_path'
        - $ref: '#/parameters/Group_in_put'        
      responses:
        '200':
                $ref: '#/responses/200_ok_only'
        '201':
                $ref: '#/responses/201_created'
        '204':
                $ref: '#/responses/204_no_content'
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '409':
                $ref: '#/responses/409_conflict'
        '500':
                $ref: '#/responses/500_internal_server_error'
    delete:
      summary: delete an existing group
      description: delete an existing group
      tags:
        - Tasks_resource
      security:
        - Api_key_teacher: []  #solo insegnante può creare
      parameters:
        - $ref: '#/parameters/Group_id_in_path'
      responses:
        '204':
                $ref: '#/responses/204_no_content'
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '405':
                $ref: '#/responses/405_method_not_allowed'
        '500':
                $ref: '#/responses/500_internal_server_error'
...
definitions:
  Error:
    properties:
      code:
        type: integer
      message:
        type: string
  
###oggetto user
...
  Exam:
    type: object
    properties:
      id:  #primary key
        type: integer 
      owner: #foreign key
        type: integer
      title:
        type: string
      description:
        type: string
      taskset:
        type: array
        items:
          $ref: '#/definitions/Task_in_exam'
      group:
        $ref: '#/definitions/Group'
      final_deadline:
        type: string
        format: date
      review_deadline:
        type: string
        format: date
  Exams:
        type: array
        items: 
                $ref: '#/definitions/Exam'
  
###oggetto task
...
  Task_in_exam:
         type: object
         properties:
                task_id: #primary key
                       type: integer
                description:
                       type: string
         
                
###oggetto group
  Group:
    type: object
    properties:
      group_id: #primary key
        type: integer
      owner_id: #foreign key
        type: integer
      name:
        type: string
      description:
        type: string
      members:
        type: array
        items:
          $ref: '#/definitions/User_in_response'
  Groups:
        type: array
        items: 
                $ref: '#/definitions/Group'
###oggetto exam_submission
  Exam_submission:
    type: object
    properties:
      exam_submission_id: #primary key
        type: integer
      referred_exam: #'foreign' key
        $ref: '#/definitions/Exam'
      submitter_id: #foreign key
        type: integer
      answers:
        type: array
        items:
          type: object
          properties:
            task_id:
              type: integer
            answer:
              type: string
      status:
        type: string
      evaluation:
        type: string
  Exam_submissions:
          type: array
          items:
                  $ref: '#/definitions/Exam_submission'  
###oggetto exam_peer_review
...
  User_in_response:
      type: object   
      properties:
        user_id: 
          type: integer
          description: the user 's id
        email:
          type: string
          description: the user 's email
        name:
          type: string
          description: the user 's name
        surname:
          type: string
          description: the user 's surname
  Users_in_response:
    type: array
    items:
      $ref: '#/definitions/User_in_response'
  User_for_post:
    type: object   
    properties:
      email:
        type: string
        description: the user 's email
      password:
        type: string
        format: password
        description: the user 's password
      name:
        type: string
        description: the user 's name
      surname:
        type: string
        description: the user 's surname
...
  Group_for_post:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      members_id:
        type: array
        items:
          type: integer
...
parameters:
  User_in_post:        
        name: user
        description: the object user
        in: body
        schema:
                $ref: '#/definitions/User_for_post'
                required:
                        - email
                        - password
                        - name
                        - surname
...
  User_id_in_path:
    name: user_id
    description: the user's id
    in: path
    type: integer
    required: true
...
  Group_in_post:
        name: group
        description: the object group
        in: body
        schema:
                $ref: '#/definitions/Group_for_post'
                required:
                        - name
                        - description
                        - members  
  Group_in_put:
        name: group
        description: the object group
        in: body
        schema:
                $ref: '#/definitions/Group_for_post'
  Group_id_in_path:
    name: group_id
    description: the id of a group
    in: path
    type: integer
    required: true    
...
responses:
...
  200_ok_only:
        description: Created  
  201_created:
        description: Created
  204_no_content:
        description: No Content
  400_bad_request:
          description: Bad Request
          schema:
                $ref: '#/definitions/Error'
  401_unauthorized:
          description: Unauthorized, missing or invalid API Key
          schema:
                $ref: '#/definitions/Error'
  404_not_found:
          description: Not Found
          schema:
                 $ref: '#/definitions/Error'
  405_method_not_allowed:
          description: Method Not Allowed
          schema:
                 $ref: '#/definitions/Error'
  409_conflict:
          description: Conflict
          schema:
                 $ref: '#/definitions/Error'
  500_internal_server_error:
          description: Internal Server Error
          schema:
                $ref: '#/definitions/Error'
         
         
#definisce apikey      
securityDefinitions:
        Api_key_student:
                type: apiKey
                in: query
                name: token
        Api_key_teacher:
                type: apiKey
                in: query
                name: tokent